+++ packages/destination-actions/src/destinations/engage/twilio/utils/message-sender.ts
@@ -6,12 +6,13 @@ import type { Payload as SmsPayload } from '../sendSms/generated-types'
 import type { Payload as WhatsappPayload } from '../sendWhatsApp/generated-types'
 import { IntegrationError, PayloadValidationError, RequestClient, RequestOptions } from '@segment/actions-core'
 import { ExecuteInput } from '@segment/actions-core'
-import { ContentTemplateResponse, ContentTemplateTypes, Profile, TwilioApiError } from './types'
-import { track, OperationContext } from './track'
-import { isDestinationActionService } from './isDestinationActionService'
-import { MessageLogger } from './MessageLogger'
-import { MessageStats } from './MessageStats'
-import { IntegrationErrorWrapper } from './IntegrationErrorWrapper'
+import { ContentTemplateResponse, ContentTemplateTypes, Profile } from './types'
+import { TwilioApiError } from '../../utils/TwilioApiError'
+import { track, OperationContext } from '../../utils/track'
+import { isDestinationActionService } from '../../utils/isDestinationActionService'
+import { MessageLogger } from '../../utils/MessageLogger'
+import { MessageStats } from '../../utils/MessageStats'
+import { IntegrationErrorWrapper } from '../../utils/IntegrationErrorWrapper'
 
 const Liquid = new LiquidJs()
