+++ packages/destination-actions/src/destinations/engage/twilio/utils/types.ts
@@ -9,21 +9,6 @@ export interface ContentTemplateResponse {
   }
 }
 
-export interface TwilioApiError extends Error {
-  response: {
-    data: {
-      code: number
-      message: string
-      more_info: string
-      status: number
-    }
-    headers?: Response['headers']
-  }
-  code?: number
-  status?: number
-  statusCode?: number
-}
-
 export interface ContentTemplateResponse {
   types: {
     [type: string]: {
