+++ packages/destination-actions/src/destinations/engage/utils/TwilioApiError.ts
@@ -1,0 +1,14 @@
+export interface TwilioApiError extends Error {
+  response: {
+    data: {
+      code: number
+      message: string
+      more_info: string
+      status: number
+    }
+    headers?: Response['headers']
+  }
+  code?: number
+  status?: number
+  statusCode?: number
+}
