+++ packages/destination-actions/src/destinations/engage/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './EngageAction'
+export * from './EngageActionPerformer'
 export * from './operationTracking'
